refactor(core): relocate QueryParams/QueryResult out of core/types
`QueryParams` and `QueryResult` describe the callback-shaped contract
each backend's `handler.ts` exposes internally — they are NOT a core
abstraction. `ChatBackend.runChatTurn` (in
`core/agent-runtime/capabilities.ts`) is the canonical surface every
consumer outside `src/backend/` talks to.

Moving them into `src/backend/shared/handler-types.ts` keeps
`core/types.ts` clean of implementation-detail shapes so the
dispatcher / cron / triggers / frontends can no longer accidentally
couple to the callback contract. `ExecuteResult` is now declared in
full rather than extending `QueryResult` (it lives at the dispatcher
layer where coupling to a backend-internal type would be wrong).

Doc strings in `agent-runtime/capabilities.ts`, `events.ts`, and
`legacy-bridge.ts` are updated so they no longer reference the
relocated types as if they were core abstractions.

Also: gitignore `.claude/` (local agent state).
