+++ .gitignore
@@ -25,3 +25,4 @@ src/__tests__/integration/stub-claude/fake-claude
 src/__tests__/integration/stub-claude/fake-claude.exe
 src/__tests__/integration/stub-claude/fake-claude.cjs
 src/__tests__/integration/stub-claude/sea-prep.blob
+.claude/
