+++ src/backend/codex/handler.ts
@@ -40,7 +40,7 @@ import type {
   McpToolCallItem,
   Usage,
 } from "@openai/codex-sdk";
-import type { QueryParams, QueryResult } from "../../core/types.js";
+import type { QueryParams, QueryResult } from "../shared/handler-types.js";
 import {
   getSession,
   incrementTurns,
