+++ src/backend/kilo/handler.ts
@@ -23,7 +23,7 @@
  */
 
 import type { KiloClient } from "@kilocode/sdk/v2";
-import type { QueryParams, QueryResult } from "../../core/types.js";
+import type { QueryParams, QueryResult } from "../shared/handler-types.js";
 import {
   getSession,
   incrementTurns,
