+++ src/backend/opencode/handler.ts
@@ -25,7 +25,7 @@
  */
 
 import type { OpencodeClient } from "@opencode-ai/sdk/v2";
-import type { QueryParams, QueryResult } from "../../core/types.js";
+import type { QueryParams, QueryResult } from "../shared/handler-types.js";
 import {
   getSession,
   incrementTurns,
